bb_scheduler: remove lock contention by switching to channels
workerBuildQueue.jobsLock is used by the Scheduler for Job enqueue/deque events
and also in GetWork() RPC for each Job completion notification. Every job completion
fires a broadcast which is caught by every other executing job, and all contend on
the same jobsLock. With this patch, we remove the dependency of the Jobs completion
notification with channels, with each job having a dedicated channel